requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
motor==3.7.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
celery[redis]==5.4.0
redis==4.6.0
boto3==1.28.30
google-api-python-client==2.99.0
stripe==6.4.0
twilio==9.5.2
email-validator==2.1.1
python-multipart==0.0.6
aiohttp==3.8.5
python-paystack==1.0.0
httpx==0.24.1